testing new iterators

+++ containers_test/srcs/vector/insert.cpp
@@ -24,13 +24,13 @@ int		main(void)
 	printSize(vct);
 
 	vct2.insert(vct2.end(), 42);
-	vct2.insert(vct2.begin(), 2, 21);
+	vct2.insert(vct2.begin(), (size_t)2, 21);
 	printSize(vct2);
 
 	vct2.insert(vct2.end() - 2, 42);
 	printSize(vct2);
 
-	vct2.insert(vct2.end(), 2, 84);
+	vct2.insert(vct2.end(), (size_t)2, 84);
 	printSize(vct2);
 
 	vct2.resize(4);
